pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "humangr-pipeline"
version = "1.0.0"
description = "Pipeline for building Human Layer MCP Server"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "pyyaml>=6.0",
    "qdrant-client>=1.7.0",
    "fastembed>=0.2.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
humangr-pipeline = "pipeline.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
